aplicado validação da conta de e-mail para o login do usuario na MainActivity.java. Criado o metodo updateUi, signIn,inicializar e chamada de eventos dos botoes na classe MainActivity.java
