+++ app/src/main/java/com/example/appsenai/FormCadastroActivity.java
@@ -72,7 +72,7 @@ public class FormCadastroActivity extends AppCompatActivity {
                         } else {
                             // If sign in fails, display a message to the user.
                             Log.w(TAG, "createUserWithEmail:failure", task.getException());
-                            Toast.makeText(FormCadastroActivity.this, "Conta já existe ou o endereço está incompleto! ",
+                            Toast.makeText(FormCadastroActivity.this, "Ocorreu um erro: "+task.getException(),
                                     Toast.LENGTH_LONG).show();
                             updateUI(null);
                         }
@@ -81,28 +81,7 @@ public class FormCadastroActivity extends AppCompatActivity {
         // [END create_user_with_email]
     }
 
-    private void signIn(String email, String password) {
-        // [START sign_in_with_email]
-        mAuth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            // Sign in success, update UI with the signed-in user's information
-                            Log.d(TAG, "signInWithEmail:success");
-                            FirebaseUser user = mAuth.getCurrentUser();
-                            updateUI(user);
-                        } else {
-                            // If sign in fails, display a message to the user.
-                            Log.w(TAG, "signInWithEmail:failure", task.getException());
-                            Toast.makeText(FormCadastroActivity.this, "Authentication failed.",
-                                    Toast.LENGTH_SHORT).show();
-                            updateUI(null);
-                        }
-                    }
-                });
-        // [END sign_in_with_email]
-    }
+
