+++ app/src/main/java/com/example/appsenai/MainActivity.java
@@ -1,29 +1,40 @@
 package com.example.appsenai;
 
 import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
+import android.widget.Toast;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 public class MainActivity extends AppCompatActivity {
     private Button botao;
+    private EditText editEmail, editSenha;
     private TextView txt_tela_cadastro;
-
+    private static final String TAG = "EmailPasswordLogin";
+    // [START declare_auth]
+    private FirebaseAuth mAuth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
+        inicializar();
         getSupportActionBar().hide();
 
-        txt_tela_cadastro = findViewById(R.id.txtTelaCadastro);
         txt_tela_cadastro.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -34,15 +45,52 @@ public class MainActivity extends AppCompatActivity {
 
 
 
-        botao = findViewById(R.id.btEntrar);
+
         botao.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(MainActivity.this, Cronograma_Usuario.class);
-                startActivity(intent);
+                signIn(editEmail.getText().toString(),editSenha.getText().toString());
+
             }
         });
 
     }
 
+    private void inicializar() {
+        editEmail = findViewById(R.id.editEmail);
+        editSenha = findViewById(R.id.editSenha);
+        txt_tela_cadastro = findViewById(R.id.txtTelaCadastro);
+        botao = findViewById(R.id.btEntrar);
+        mAuth = FirebaseAuth.getInstance();
+    }
+
+    private void signIn(String email, String password) {
+        // [START sign_in_with_email]
+        mAuth.signInWithEmailAndPassword(email, password)
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            // Sign in success, update UI with the signed-in user's information
+                            Log.d(TAG, "signInWithEmail:success");
+                            FirebaseUser user = mAuth.getCurrentUser();
+                            updateUI(user);
+                        } else {
+                            // If sign in fails, display a message to the user.
+                            Log.w(TAG, "signInWithEmail:failure", task.getException());
+                            Toast.makeText(MainActivity.this, "Autenticação Falhou.",
+                                    Toast.LENGTH_SHORT).show();
+                            updateUI(null);
+                        }
+                    }
+                });
+        // [END sign_in_with_email]
+    }
+
+    private void updateUI(FirebaseUser user) {
+        if(user!=null){
+            Intent intent = new Intent(MainActivity.this, Cronograma_Usuario.class);
+            startActivity(intent);
+        }
+    }
 }
