+++ app/src/main/res/layout/activity_form_cadastro.xml
@@ -57,7 +57,6 @@
         android:layout_marginEnd="8dp"
         android:background="@drawable/edittext"
         android:hint="Digite seu Email"
-        android:text="meutestedeemial@gmail.com"
         android:inputType="text"
         android:maxLines="1"
         android:padding="15dp"
@@ -77,7 +76,6 @@
         android:hint="Digite sua senha"
         android:inputType="textPassword"
         android:maxLines="1"
-        android:text="112233"
         android:padding="15dp"
         android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
